♻️[Refactor] KOPIS 날짜 유틸 date.ts로 분리 및 공연 탐색 기본 필터 개선
- formatKopisDate, toKopisDate, addMonths, addYears → date.ts로 이동
- 공연 탐색 기본 상태 필터를 공연예정(01)으로 변경
- 검색 시 상태 전체 초기화 및 기간 2000-01-01부터로 확장
