+++ src/shared/utils/date.ts
@@ -34,3 +34,32 @@ export function localDateToRfc3339NoonUtc(localDate: string): string {
   // ✅ 정오(12:00:00.000Z)로 고정
   return `${localDate}T12:00:00.000Z`;
 }
+
+// ─────────────────────────────────────────────
+// KOPIS 날짜 유틸
+// ─────────────────────────────────────────────
+
+/** Date → KOPIS API 파라미터 포맷 "YYYYMMDD" */
+export function formatKopisDate(date: Date): string {
+  const y = date.getFullYear();
+  const m = String(date.getMonth() + 1).padStart(2, "0");
+  const d = String(date.getDate()).padStart(2, "0");
+  return `${y}${m}${d}`;
+}
+
+/** "YYYY-MM-DD" → "YYYYMMDD" */
+export function toKopisDate(v: string): string {
+  return v.replace(/-/g, "").slice(0, 8);
+}
+
+export function addMonths(date: Date, months: number): Date {
+  const d = new Date(date);
+  d.setMonth(d.getMonth() + months);
+  return d;
+}
+
+export function addYears(date: Date, years: number): Date {
+  const d = new Date(date);
+  d.setFullYear(d.getFullYear() + years);
+  return d;
+}
